More tuning options + HW latency tester support for slow render
New arguments for test-slow-render.html:

cubeScale: Cube size multiplier for fine-grained latency tuning.

renderScale: Adjust the render resolution as a multiplier applied on top of the default renderWidth/renderHeight. Default 1.0 means to use them as-is.

latencyPatch: if true, draw a square in the right eye view with brightness proportional to the headset's angular velocity. This is intended to support hardware latency testing.